+++ src/main/java/com/dao/AddressBookJoinDao.java
@@ -1,16 +1,15 @@
 package com.dao;
 
 import com.entity.AddressBookJoinEntity;
-import java.math.BigDecimal;
 
 public interface AddressBookJoinDao {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(AddressBookJoinEntity record);
 
     int insertSelective(AddressBookJoinEntity record);
 
-    AddressBookJoinEntity selectByPrimaryKey(BigDecimal id);
+    AddressBookJoinEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(AddressBookJoinEntity record);
