+++ src/main/java/com/dao/SysConfigDao.java
@@ -1,16 +1,15 @@
 package com.dao;
 
 import com.entity.SysConfigEntity;
-import java.math.BigDecimal;
 
 public interface SysConfigDao {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SysConfigEntity record);
 
     int insertSelective(SysConfigEntity record);
 
-    SysConfigEntity selectByPrimaryKey(BigDecimal id);
+    SysConfigEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(SysConfigEntity record);
