+++ src/main/java/com/dao/SysRolePermissionDao.java
@@ -1,16 +1,15 @@
 package com.dao;
 
 import com.entity.SysRolePermissionEntity;
-import java.math.BigDecimal;
 
 public interface SysRolePermissionDao {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SysRolePermissionEntity record);
 
     int insertSelective(SysRolePermissionEntity record);
 
-    SysRolePermissionEntity selectByPrimaryKey(BigDecimal id);
+    SysRolePermissionEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(SysRolePermissionEntity record);
