+++ src/main/java/com/dao/SysUserDao.java
@@ -1,16 +1,15 @@
 package com.dao;
 
 import com.entity.SysUserEntity;
-import java.math.BigDecimal;
 
 public interface SysUserDao {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SysUserEntity record);
 
     int insertSelective(SysUserEntity record);
 
-    SysUserEntity selectByPrimaryKey(BigDecimal id);
+    SysUserEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(SysUserEntity record);
