+++ src/main/java/com/dao/SysUserInfoDao.java
@@ -1,16 +1,15 @@
 package com.dao;
 
 import com.entity.SysUserInfoEntity;
-import java.math.BigDecimal;
 
 public interface SysUserInfoDao {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SysUserInfoEntity record);
 
     int insertSelective(SysUserInfoEntity record);
 
-    SysUserInfoEntity selectByPrimaryKey(BigDecimal id);
+    SysUserInfoEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(SysUserInfoEntity record);
