+++ src/main/java/com/entity/ActivityEntity.java
@@ -1,12 +1,11 @@
 package com.entity;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class ActivityEntity {
-    private BigDecimal id;
+    private Long id;
 
-    private BigDecimal userId;
+    private Long userId;
 
     private String title;
 
@@ -16,7 +15,7 @@ public class ActivityEntity {
 
     private Date targetDate;
 
-    private Short status;
+    private Byte status;
 
     private Date endDate;
 
@@ -26,19 +25,19 @@ public class ActivityEntity {
 
     private String password;
 
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public BigDecimal getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigDecimal userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -74,11 +73,11 @@ public class ActivityEntity {
         this.targetDate = targetDate;
     }
 
-    public Short getStatus() {
+    public Byte getStatus() {
         return status;
     }
 
-    public void setStatus(Short status) {
+    public void setStatus(Byte status) {
         this.status = status;
     }
