+++ src/main/java/com/entity/AddressBookEntity.java
@@ -1,12 +1,11 @@
 package com.entity;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class AddressBookEntity {
-    private BigDecimal id;
+    private Long id;
 
-    private BigDecimal userId;
+    private Long userId;
 
     private String title;
 
@@ -16,23 +15,23 @@ public class AddressBookEntity {
 
     private Date updateDate;
 
-    private Short status;
+    private Byte status;
 
     private String password;
 
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public BigDecimal getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigDecimal userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -68,11 +67,11 @@ public class AddressBookEntity {
         this.updateDate = updateDate;
     }
 
-    public Short getStatus() {
+    public Byte getStatus() {
         return status;
     }
 
-    public void setStatus(Short status) {
+    public void setStatus(Byte status) {
         this.status = status;
     }
