+++ src/main/java/com/entity/SysPermissionEntity.java
@@ -1,17 +1,15 @@
 package com.entity;
 
-import java.math.BigDecimal;
-
 public class SysPermissionEntity {
-    private BigDecimal id;
+    private Long id;
 
     private String name;
 
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
