+++ src/main/java/com/entity/VoteEntity.java
@@ -1,18 +1,17 @@
 package com.entity;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class VoteEntity {
-    private BigDecimal id;
+    private Long id;
 
     private String title;
 
     private String content;
 
-    private Short status;
+    private Byte status;
 
-    private BigDecimal userId;
+    private Long userId;
 
     private String password;
 
@@ -22,11 +21,11 @@ public class VoteEntity {
 
     private Date endDate;
 
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -46,19 +45,19 @@ public class VoteEntity {
         this.content = content == null ? null : content.trim();
     }
 
-    public Short getStatus() {
+    public Byte getStatus() {
         return status;
     }
 
-    public void setStatus(Short status) {
+    public void setStatus(Byte status) {
         this.status = status;
     }
 
-    public BigDecimal getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigDecimal userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
